Avoid time_t printf format warnings by casting to long.
Clang warns on using %d to format the struct timeval tv_sec member, as
it isn't necessarily an int. While long isn't guaranteed to be long
enough either, let's use that anyway. It's good enough.

+++ multibot.c
@@ -399,7 +399,7 @@ char *timestamp()
 {
     struct timeval tv;
     if (gettimeofday(&tv, NULL) == 0) {
-        sprintf(tsbuf, "%d %d", tv.tv_sec, tv.tv_usec);
+        sprintf(tsbuf, "%ld %ld", (long) tv.tv_sec, (long) tv.tv_usec);
     } else {
         sprintf(tsbuf, "? ?");
     }
